+++ src/language/stella-module.ts
@@ -20,6 +20,7 @@ import { CompletionProvider } from "./lsp/completion-provider.js";
 import { StellaScopeProvider } from "./references/scope-provider.js";
 import { StellaScopeComputation } from "./references/scope-computation.js";
 import { HoverProvider } from "./lsp/hover-provider.js";
+import { StellaCodeActionProvider } from "./lsp/code-action-provider.js";
 
 /**
  * Declaration of custom services - add your own service classes here.
@@ -52,6 +53,7 @@ export const StellaModule: Module<
     SemanticTokenProvider: (services) => new SemanticTokenProvider(services),
     CompletionProvider: (services) => new CompletionProvider(services),
     HoverProvider: (services) => new HoverProvider(services),
+    CodeActionProvider: (services) => new StellaCodeActionProvider(services),
   },
   references: {
     ScopeProvider: (services) => new StellaScopeProvider(services),